itertools + _contextvars + _abc + _functools: extract into module/{itertools,_contextvars,_abc,_functools}/
importing.rs shrinks by 463 lines.
