+++ pyre/pyre-interpreter/src/importing.rs
@@ -112,11 +112,17 @@ pub fn install_builtin_modules() {
     // Minimal C-extension stubs required for stdlib import chains.
     // PyPy: these are all implemented as mixed modules under pypy/module/.
     register_builtin_module("_weakref", crate::module::_weakref::moduledef::init);
-    register_builtin_module("_abc", init_abc);
-    register_builtin_module("_functools", init_functools);
+    register_builtin_module("_abc", crate::module::_abc::moduledef::init);
+    register_builtin_module(
+        "_functools",
+        crate::module::_functools::moduledef::init,
+    );
     register_builtin_module("_thread", init_thread);
-    register_builtin_module("itertools", init_itertools);
-    register_builtin_module("_contextvars", init_contextvars);
+    register_builtin_module("itertools", crate::module::itertools::moduledef::init);
+    register_builtin_module(
+        "_contextvars",
+        crate::module::_contextvars::moduledef::init,
+    );
     register_builtin_module("copyreg", init_copyreg);
     register_builtin_module("_codecs", init_codecs);
     register_builtin_module("posix", init_posix);
@@ -245,473 +251,6 @@ fn stat_result_type() -> PyObjectRef {
             tp
         })
     })
-}
-
-/// itertools stub
-fn init_itertools(ns: &mut DictStorage) {
-    // chain(*iterables) → flat iterator
-    crate::dict_storage_store(
-        ns,
-        "chain",
-        crate::make_builtin_function("chain", |args| {
-            let mut items = Vec::new();
-            for &arg in args {
-                items.extend(crate::builtins::collect_iterable(arg)?);
-            }
-            let n = items.len();
-            let list = pyre_object::w_list_new(items);
-            Ok(pyre_object::w_seq_iter_new(list, n))
-        }),
-    );
-    // starmap stub
-    crate::dict_storage_store(
-        ns,
-        "starmap",
-        crate::make_builtin_function_with_arity(
-            "starmap",
-            |_| Ok(pyre_object::w_list_new(vec![])),
-            2,
-        ),
-    );
-    // count(start=0, step=1) — PyPy: W_Count___new__
-    //
-    //     def W_Count___new__(space, w_subtype, w_start=0, w_step=1):
-    //         return W_Count(space, w_start, w_step)
-    crate::dict_storage_store(
-        ns,
-        "count",
-        crate::make_builtin_function("count", |args| {
-            let w_start = args.first().copied().unwrap_or(pyre_object::w_int_new(0));
-            let w_step = args.get(1).copied().unwrap_or(pyre_object::w_int_new(1));
-            Ok(pyre_object::itertoolsmodule::w_count_new(w_start, w_step))
-        }),
-    );
-    // repeat(obj, times=None) — PyPy: W_Repeat___new__
-    //
-    //     def W_Repeat___new__(space, w_subtype, w_obj, w_times=None):
-    //         return W_Repeat(space, w_obj, w_times)
-    crate::dict_storage_store(
-        ns,
-        "repeat",
-        crate::make_builtin_function("repeat", |args| {
-            if args.is_empty() {
-                return Err(crate::PyError::type_error(
-                    "repeat() missing 'object' argument",
-                ));
-            }
-            let w_obj = args[0];
-            let w_times = if args.len() >= 2 {
-                unsafe {
-                    if pyre_object::is_int(args[1]) {
-                        Some(pyre_object::w_int_get_value(args[1]))
-                    } else {
-                        None
-                    }
-                }
-            } else {
-                None
-            };
-            Ok(pyre_object::itertoolsmodule::w_repeat_new(w_obj, w_times))
-        }),
-    );
-    // islice
-    crate::dict_storage_store(
-        ns,
-        "islice",
-        crate::make_builtin_function("islice", |_| Ok(pyre_object::w_list_new(vec![]))),
-    );
-    // groupby
-    crate::dict_storage_store(
-        ns,
-        "groupby",
-        crate::make_builtin_function("groupby", |_| Ok(pyre_object::w_none())),
-    );
-    // permutations(iterable, r=None) — PyPy: pypy/module/itertools/interp_itertools.py
-    crate::dict_storage_store(
-        ns,
-        "permutations",
-        crate::make_builtin_function("permutations", |args| {
-            if args.is_empty() {
-                return Ok(pyre_object::w_list_new(vec![]));
-            }
-            let pool = crate::builtins::collect_iterable(args[0])?;
-            let n = pool.len();
-            let r = if args.len() >= 2 {
-                unsafe {
-                    if pyre_object::is_int(args[1]) {
-                        pyre_object::w_int_get_value(args[1]) as usize
-                    } else {
-                        n
-                    }
-                }
-            } else {
-                n
-            };
-            if r > n {
-                return Ok(pyre_object::w_list_new(vec![]));
-            }
-            // Heap/Lehmer would be clearer; use a recursive closure-free helper.
-            fn perms(
-                pool: &[pyre_object::PyObjectRef],
-                r: usize,
-            ) -> Vec<Vec<pyre_object::PyObjectRef>> {
-                if r == 0 {
-                    return vec![vec![]];
-                }
-                let mut out = Vec::new();
-                for i in 0..pool.len() {
-                    let mut rest: Vec<_> = pool.to_vec();
-                    let head = rest.remove(i);
-                    for mut tail in perms(&rest, r - 1) {
-                        let mut v = vec![head];
-                        v.append(&mut tail);
-                        out.push(v);
-                    }
-                }
-                out
-            }
-            let all = perms(&pool, r);
-            let tuples: Vec<_> = all.into_iter().map(pyre_object::w_tuple_new).collect();
-            let n = tuples.len();
-            let list = pyre_object::w_list_new(tuples);
-            Ok(pyre_object::w_seq_iter_new(list, n))
-        }),
-    );
-    // combinations(iterable, r)
-    crate::dict_storage_store(
-        ns,
-        "combinations",
-        crate::make_builtin_function_with_arity(
-            "combinations",
-            |args| {
-                if args.len() < 2 {
-                    return Ok(pyre_object::w_list_new(vec![]));
-                }
-                let pool = crate::builtins::collect_iterable(args[0])?;
-                let r = unsafe { pyre_object::w_int_get_value(args[1]) as usize };
-                if r > pool.len() {
-                    return Ok(pyre_object::w_list_new(vec![]));
-                }
-                fn combs(
-                    pool: &[pyre_object::PyObjectRef],
-                    r: usize,
-                    start: usize,
-                ) -> Vec<Vec<pyre_object::PyObjectRef>> {
-                    if r == 0 {
-                        return vec![vec![]];
-                    }
-                    let mut out = Vec::new();
-                    for i in start..pool.len() {
-                        for mut tail in combs(pool, r - 1, i + 1) {
-                            let mut v = vec![pool[i]];
-                            v.append(&mut tail);
-                            out.push(v);
-                        }
-                    }
-                    out
-                }
-                let all = combs(&pool, r, 0);
-                let tuples: Vec<_> = all.into_iter().map(pyre_object::w_tuple_new).collect();
-                let n = tuples.len();
-                let list = pyre_object::w_list_new(tuples);
-                Ok(pyre_object::w_seq_iter_new(list, n))
-            },
-            2,
-        ),
-    );
-    // product(*iterables, repeat=1)
-    crate::dict_storage_store(
-        ns,
-        "product",
-        crate::make_builtin_function("product", |args| {
-            let pools: Vec<Vec<_>> = args
-                .iter()
-                .map(|&a| crate::builtins::collect_iterable(a))
-                .collect::<Result<_, _>>()?;
-            let mut result: Vec<Vec<pyre_object::PyObjectRef>> = vec![vec![]];
-            for pool in &pools {
-                let mut new_result = Vec::with_capacity(result.len() * pool.len());
-                for existing in &result {
-                    for &item in pool {
-                        let mut v = existing.clone();
-                        v.push(item);
-                        new_result.push(v);
-                    }
-                }
-                result = new_result;
-            }
-            let tuples: Vec<_> = result.into_iter().map(pyre_object::w_tuple_new).collect();
-            let n = tuples.len();
-            let list = pyre_object::w_list_new(tuples);
-            Ok(pyre_object::w_seq_iter_new(list, n))
-        }),
-    );
-    // zip_longest(*iterables, fillvalue=None) — interp_itertools.py
-    // W_ZipLongest. CALL_KW packs `fillvalue` into the trailing
-    // `__pyre_kw__` dict (`call.rs:727-744`); strip it before
-    // collecting the iterable pools so the kwarg doesn't surface as
-    // an extra positional pool.
-    crate::dict_storage_store(
-        ns,
-        "zip_longest",
-        crate::make_builtin_function("zip_longest", |args| {
-            let (positional, kwargs) = crate::builtins::split_builtin_kwargs(args);
-            // `pypy/module/itertools/interp_itertools.py:685` —
-            // W_ZipLongest's `unwrap_spec` only knows about
-            // `fillvalue`; any other keyword raises TypeError at the
-            // gateway.  Pyre's flat builtin ABI has to enforce this
-            // by hand.
-            crate::builtins::kwarg_reject_unknown(kwargs, &["fillvalue"], "zip_longest")?;
-            let fill =
-                crate::builtins::kwarg_get(kwargs, "fillvalue").unwrap_or_else(pyre_object::w_none);
-            let pools: Vec<Vec<_>> = positional
-                .iter()
-                .map(|&a| crate::builtins::collect_iterable(a))
-                .collect::<Result<_, _>>()?;
-            let max_len = pools.iter().map(|p| p.len()).max().unwrap_or(0);
-            let mut tuples = Vec::with_capacity(max_len);
-            for i in 0..max_len {
-                let row: Vec<_> = pools
-                    .iter()
-                    .map(|p| if i < p.len() { p[i] } else { fill })
-                    .collect();
-                tuples.push(pyre_object::w_tuple_new(row));
-            }
-            let n = tuples.len();
-            let list = pyre_object::w_list_new(tuples);
-            Ok(pyre_object::w_seq_iter_new(list, n))
-        }),
-    );
-    // accumulate(iterable) — sums only, PyPy interp_itertools W_Accumulate.
-    crate::dict_storage_store(
-        ns,
-        "accumulate",
-        crate::make_builtin_function("accumulate", |args| {
-            if args.is_empty() {
-                return Ok(pyre_object::w_list_new(vec![]));
-            }
-            let items = crate::builtins::collect_iterable(args[0])?;
-            let mut out = Vec::with_capacity(items.len());
-            let mut acc: Option<pyre_object::PyObjectRef> = None;
-            for item in items {
-                acc = Some(match acc {
-                    None => item,
-                    Some(prev) => crate::baseobjspace::add(prev, item)?,
-                });
-                out.push(acc.unwrap());
-            }
-            let n = out.len();
-            let list = pyre_object::w_list_new(out);
-            Ok(pyre_object::w_seq_iter_new(list, n))
-        }),
-    );
-    // compress(data, selectors)
-    crate::dict_storage_store(
-        ns,
-        "compress",
-        crate::make_builtin_function_with_arity(
-            "compress",
-            |args| {
-                if args.len() < 2 {
-                    return Ok(pyre_object::w_list_new(vec![]));
-                }
-                let data = crate::builtins::collect_iterable(args[0])?;
-                let selectors = crate::builtins::collect_iterable(args[1])?;
-                let mut out = Vec::new();
-                for (d, s) in data.iter().zip(selectors.iter()) {
-                    if crate::baseobjspace::is_true(*s) {
-                        out.push(*d);
-                    }
-                }
-                let n = out.len();
-                let list = pyre_object::w_list_new(out);
-                Ok(pyre_object::w_seq_iter_new(list, n))
-            },
-            2,
-        ),
-    );
-}
-
-/// _contextvars stub
-fn init_contextvars(ns: &mut DictStorage) {
-    // ContextVar(name, *, default=_MISSING) — context variable
-    crate::dict_storage_store(
-        ns,
-        "ContextVar",
-        crate::make_builtin_function("ContextVar", |args| {
-            // Return stub object with get/set methods
-            let obj = pyre_object::w_instance_new(crate::typedef::w_object());
-            if !args.is_empty() {
-                let _ = crate::baseobjspace::setattr(obj, "name", args[0]);
-            }
-            // get() returns default or raises LookupError
-            let _ = crate::baseobjspace::setattr(
-                obj,
-                "get",
-                crate::make_builtin_function("get", |args| {
-                    // Return default if provided
-                    if args.len() > 1 {
-                        Ok(args[1])
-                    } else {
-                        Ok(pyre_object::w_none())
-                    }
-                }),
-            );
-            let _ = crate::baseobjspace::setattr(
-                obj,
-                "set",
-                crate::make_builtin_function_with_arity("set", |_| Ok(pyre_object::w_none()), 2),
-            );
-            Ok(obj)
-        }),
-    );
-    crate::dict_storage_store(
-        ns,
-        "Context",
-        crate::make_builtin_function_with_arity("Context", |_| Ok(pyre_object::w_none()), 0),
-    );
-    crate::dict_storage_store(
-        ns,
-        "Token",
-        crate::make_builtin_function_with_arity("Token", |_| Ok(pyre_object::w_none()), 0),
-    );
-    crate::dict_storage_store(
-        ns,
-        "copy_context",
-        crate::make_builtin_function_with_arity("copy_context", |_| Ok(pyre_object::w_none()), 0),
-    );
-}
-
-/// _abc stub — PyPy: pypy/module/_abc/
-fn init_abc(ns: &mut DictStorage) {
-    crate::dict_storage_store(
-        ns,
-        "get_cache_token",
-        crate::make_builtin_function_with_arity(
-            "get_cache_token",
-            |_| Ok(pyre_object::w_int_new(0)),
-            0,
-        ),
-    );
-    crate::dict_storage_store(
-        ns,
-        "_abc_init",
-        crate::make_builtin_function_with_arity("_abc_init", |_| Ok(pyre_object::w_none()), 1),
-    );
-    crate::dict_storage_store(
-        ns,
-        "_abc_register",
-        crate::make_builtin_function_with_arity("_abc_register", |_| Ok(pyre_object::w_none()), 2),
-    );
-    // _abc_instancecheck(cls, instance) — CPython: Modules/_abc.c _abc__abc_instancecheck.
-    //
-    // ABCMeta.__instancecheck__ (abc.py:119) delegates here. The canonical
-    // behaviour: walk type(instance).__mro__ looking for cls (direct
-    // subclass), then consult cls._abc_registry for virtual subclasses
-    // registered via `cls.register(subclass)`. Our previous stub
-    // unconditionally returned False, which broke
-    // `isinstance(Fraction(1,2), numbers.Rational)`.
-    crate::dict_storage_store(
-        ns,
-        "_abc_instancecheck",
-        crate::make_builtin_function_with_arity(
-            "_abc_instancecheck",
-            |args| {
-                if args.len() < 2 {
-                    return Ok(pyre_object::w_bool_from(false));
-                }
-                let cls = args[0];
-                let instance = args[1];
-                unsafe {
-                    Ok(pyre_object::w_bool_from(crate::baseobjspace::isinstance_w(
-                        instance, cls,
-                    )))
-                }
-            },
-            2,
-        ),
-    );
-    // _abc_subclasscheck(cls, subclass) — CPython: Modules/_abc.c _abc__abc_subclasscheck.
-    crate::dict_storage_store(
-        ns,
-        "_abc_subclasscheck",
-        crate::make_builtin_function_with_arity(
-            "_abc_subclasscheck",
-            |args| {
-                if args.len() < 2 {
-                    return Ok(pyre_object::w_bool_from(false));
-                }
-                let cls = args[0];
-                let subclass = args[1];
-                unsafe {
-                    // Walk subclass.__mro__ looking for cls.
-                    let mro_ptr = pyre_object::w_type_get_mro(subclass);
-                    if !mro_ptr.is_null() {
-                        for &t in &*mro_ptr {
-                            if std::ptr::eq(t, cls) {
-                                return Ok(pyre_object::w_bool_from(true));
-                            }
-                        }
-                    }
-                }
-                Ok(pyre_object::w_bool_from(false))
-            },
-            2,
-        ),
-    );
-    crate::dict_storage_store(
-        ns,
-        "_get_dump",
-        crate::make_builtin_function_with_arity(
-            "_get_dump",
-            |_| Ok(pyre_object::w_tuple_new(vec![])),
-            1,
-        ),
-    );
-    crate::dict_storage_store(
-        ns,
-        "_reset_registry",
-        crate::make_builtin_function_with_arity(
-            "_reset_registry",
-            |_| Ok(pyre_object::w_none()),
-            1,
-        ),
-    );
-    crate::dict_storage_store(
-        ns,
-        "_reset_caches",
-        crate::make_builtin_function_with_arity("_reset_caches", |_| Ok(pyre_object::w_none()), 1),
-    );
-}
-
-/// _functools stub
-fn init_functools(ns: &mut DictStorage) {
-    crate::dict_storage_store(
-        ns,
-        "reduce",
-        crate::make_builtin_function("reduce", |_| {
-            Err(crate::PyError::type_error("reduce not implemented"))
-        }),
-    );
-    // functools.cmp_to_key(cmp) — returns a callable that wraps a value in
-    // an opaque key. For sorting str / int / tuple of those (the only paths
-    // pyre's stdlib actually exercises), the items are already comparable,
-    // so an identity key gives the same ordering as `cmp(a, b)` would.
-    crate::dict_storage_store(
-        ns,
-        "cmp_to_key",
-        crate::make_builtin_function_with_arity(
-            "cmp_to_key",
-            |_args| {
-                Ok(crate::make_builtin_function_with_arity(
-                    "cmp_to_key.K",
-                    |args| Ok(args.first().copied().unwrap_or(pyre_object::w_none())),
-                    1,
-                ))
-            },
-            1,
-        ),
-    );
 }
 
 /// Lock methods — PyPy: pypy/module/thread/os_lock.py W_Lock / W_RLock
