+++ pyre/pyre-interpreter/src/module/_abc/mod.rs
@@ -1,0 +1,4 @@
+//! _abc module — PyPy: pypy/module/_abc/
+
+pub mod interp_abc;
+pub mod moduledef;
