+++ pyre/pyre-interpreter/src/module/_contextvars/mod.rs
@@ -1,0 +1,6 @@
+//! _contextvars module — PyPy: pypy/module/_contextvars/
+//!
+//! Stub providing ContextVar/Context/Token shells.
+
+pub mod interp_contextvars;
+pub mod moduledef;
