+++ pyre/pyre-interpreter/src/module/_functools/mod.rs
@@ -1,0 +1,4 @@
+//! _functools module — PyPy: pypy/module/_functools/
+
+pub mod interp_functools;
+pub mod moduledef;
