+++ pyre/pyre-interpreter/src/module/_functools/moduledef.rs
@@ -1,0 +1,7 @@
+//! _functools module definition — PyPy: pypy/module/_functools/moduledef.py
+
+use crate::DictStorage;
+
+pub fn init(ns: &mut DictStorage) {
+    super::interp_functools::register_module(ns);
+}
