+++ pyre/pyre-interpreter/src/module/itertools/mod.rs
@@ -1,0 +1,4 @@
+//! itertools module — PyPy: pypy/module/itertools/
+
+pub mod interp_itertools;
+pub mod moduledef;
