+++ pyre/pyre-interpreter/src/module/itertools/moduledef.rs
@@ -1,0 +1,7 @@
+//! itertools module definition — PyPy: pypy/module/itertools/moduledef.py
+
+use crate::DictStorage;
+
+pub fn init(ns: &mut DictStorage) {
+    super::interp_itertools::register_module(ns);
+}
